requirements.txt
# O-Zone MVP Dependencies

# Web Framework
streamlit>=1.30.0

# AWS SDK for Bedrock
boto3>=1.34.0

# HTTP Client for OpenAQ API
httpx>=0.27.0

# Data manipulation and visualization
pandas>=2.1.0
plotly>=5.18.0

# Globe visualization
pydeck>=0.8.0
folium>=0.15.0
streamlit-folium>=0.15.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.92.0

# Mocking for tests
responses>=0.24.0
moto>=4.2.0
freezegun>=1.4.0
